requirements.txt
torch
tensorly
numpys